enable braa/blraa, move test unit from sample_arm64 to test_arm64
